Fixed package names, refactored code to support.
