+++ producer/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.debugnotes.example.contract</groupId>
+    <groupId>com.debugnotes.example</groupId>
     <artifactId>producer</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
@@ -68,7 +68,7 @@
                 <version>${spring-cloud-contract-maven-plugin.version}</version>
                 <extensions>true</extensions>
                 <configuration>
-                    <baseClassForTests>com.debugnotes.example.contract.producer.ContractVerifierBase</baseClassForTests>
+                    <baseClassForTests>com.debugnotes.example.producer.ContractVerifierBase</baseClassForTests>
                     <!--<packageWithBaseClasses>com.debugnotes.example.contract.producer</packageWithBaseClasses>-->
                 </configuration>
             </plugin>
