+++ producer/src/main/java/com/debugnotes/example/producer/ProducerApplication.java
@@ -1,4 +1,4 @@
-package com.debugnotes.example.contract.producer;
+package com.debugnotes.example.producer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
