+++ producer/src/main/java/com/debugnotes/example/producer/model/ProfileData.java
@@ -1,4 +1,4 @@
-package com.debugnotes.example.contract.producer.model;
+package com.debugnotes.example.producer.model;
 
 public class ProfileData {
