+++ producer/src/test/java/com/debugnotes/example/producer/ContractVerifierBase.java
@@ -1,4 +1,4 @@
-package com.debugnotes.example.contract.producer;
+package com.debugnotes.example.producer;
 
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import org.junit.Before;
